# go-ecommerce-api/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "A RESTful E-commerce API built with Go and Gin",
        "title": "E-commerce API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "Authenticate user and return access and refresh tokens",
                "tags": [
                    "auth"
                ],
                "summary": "Login user",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/handlers.LoginPayload"
                            }
                        }
                    },
                    "description": "Login credentials",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.LoginResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "get": {
                "description": "Generate new access and refresh tokens using refresh token",
                "tags": [
                    "auth"
                ],
                "summary": "Refresh access token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.RefreshTokenResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Create a new user account",
                "tags": [
                    "auth"
                ],
                "summary": "Register new user",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/handlers.RegisterPayload"
                            }
                        }
                    },
                    "description": "User registration data",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.RegisterResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "//localhost:8080"
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "description": "Type \"Bearer\" followed by a space and JWT token.",
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        },
        "schemas": {
            "handlers.LoginPayload": {
                "type": "object",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "minLength": 3,
                        "example": "test@example.com"
                    },
                    "password": {
                        "type": "string",
                        "minLength": 8,
                        "example": "password"
                    }
                }
            },
            "handlers.RegisterPayload": {
                "type": "object",
                "required": [
                    "email",
                    "name",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "minLength": 3,
                        "example": "test@example.com"
                    },
                    "name": {
                        "type": "string",
                        "minLength": 2,
                        "example": "John Doe"
                    },
                    "password": {
                        "type": "string",
                        "minLength": 8,
                        "example": "password"
                    }
                }
            },
...
            "models.ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "models.LoginResponse": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "tokens": {
                        "type": "object",
                        "additionalProperties": true
                    },
                    "user": {
                        "type": "object",
                        "additionalProperties": true
                    }
                }
            },
...
            "models.RefreshTokenResponse": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "tokens": {
                        "type": "object",
                        "additionalProperties": true
                    }
                }
            },
            "models.RegisterResponse": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "tokens": {
                        "type": "object",
                        "additionalProperties": true
                    },
                    "user": {
                        "type": "object",
                        "additionalProperties": true
                    }
                }
            },
...
        }
    }
}